+++ CHANGELOG.rst
@@ -7,7 +7,9 @@ This document describes changes between each past release.
 10.2.0 (unreleased)
 ===================
 
-- Nothing changed yet.
+**New feature**
+
+- Created new method on client to get paginated records ``get_paginated_records``. (#175)
 
 
 10.1.1 (2018-11-13)